requirements.txt
python-telegram-bot==22.2
pywhispercpp
python-dotenv
asyncio
aiofiles